Fix ISAN warning reported by HighbdIntraPred unit test
This patch fixes the runtime error caused by implicit
int to uint32_t conversion reported by integer sanitizer.

+++ aom_dsp/x86/highbd_intrapred_sse2.c
@@ -827,7 +827,7 @@ void aom_highbd_dc_predictor_4x8_sse2(uint16_t *dst, ptrdiff_t stride,
   const __m128i sum_above = dc_sum_4(above);
   const __m128i sum_left = dc_sum_8(left);
   const __m128i sum = _mm_add_epi16(sum_above, sum_left);
-  uint32_t sum32 = _mm_cvtsi128_si32(sum);
+  uint32_t sum32 = (uint32_t)_mm_cvtsi128_si32(sum);
   sum32 >>= 16;
 #if CONFIG_DC_DIV_UNIFY
   int16_t shift = 0;
@@ -838,7 +838,7 @@ void aom_highbd_dc_predictor_4x8_sse2(uint16_t *dst, ptrdiff_t stride,
   sum32 += 6;
   sum32 /= 12;
 #endif  // CONFIG_DC_DIV_UNIFY
-  const __m128i row = _mm_set1_epi16((uint16_t)sum32);
+  const __m128i row = _mm_set1_epi16((int16_t)sum32);
   int i;
   for (i = 0; i < 4; ++i) {
     _mm_storel_epi64((__m128i *)dst, row);
@@ -857,7 +857,7 @@ void aom_highbd_dc_predictor_8x4_sse2(uint16_t *dst, ptrdiff_t stride,
   const __m128i sum_left = dc_sum_4(left);
   const __m128i sum_above = dc_sum_8(above);
   const __m128i sum = _mm_add_epi16(sum_above, sum_left);
-  uint32_t sum32 = _mm_cvtsi128_si32(sum);
+  uint32_t sum32 = (uint32_t)_mm_cvtsi128_si32(sum);
   sum32 >>= 16;
 #if CONFIG_DC_DIV_UNIFY
   int16_t shift = 0;
@@ -868,7 +868,7 @@ void aom_highbd_dc_predictor_8x4_sse2(uint16_t *dst, ptrdiff_t stride,
   sum32 += 6;
   sum32 /= 12;
 #endif  // CONFIG_DC_DIV_UNIFY
-  const __m128i row = _mm_set1_epi16((uint16_t)sum32);
+  const __m128i row = _mm_set1_epi16((int16_t)sum32);
 
   _mm_storeu_si128((__m128i *)dst, row);
   dst += stride;
@@ -891,7 +891,7 @@ void aom_highbd_dc_predictor_8x16_sse2(uint16_t *dst, ptrdiff_t stride,
   sum_left = _mm_unpacklo_epi16(sum_left, zero);
   sum_above = _mm_unpacklo_epi16(sum_above, zero);
   const __m128i sum = _mm_add_epi32(sum_left, sum_above);
-  uint32_t sum32 = _mm_cvtsi128_si32(sum);
+  uint32_t sum32 = (uint32_t)_mm_cvtsi128_si32(sum);
 #if CONFIG_DC_DIV_UNIFY
   int16_t shift = 0;
   uint16_t scale = resolve_divisor_32(24, &shift);
@@ -901,7 +901,7 @@ void aom_highbd_dc_predictor_8x16_sse2(uint16_t *dst, ptrdiff_t stride,
   sum32 += 12;
   sum32 /= 24;
 #endif  // CONFIG_DC_DIV_UNIFY
-  const __m128i row = _mm_set1_epi16((uint16_t)sum32);
+  const __m128i row = _mm_set1_epi16((int16_t)sum32);
   int i;
   for (i = 0; i < 4; ++i) {
     _mm_storeu_si128((__m128i *)dst, row);
@@ -927,7 +927,7 @@ void aom_highbd_dc_predictor_16x8_sse2(uint16_t *dst, ptrdiff_t stride,
   sum_left = _mm_unpacklo_epi16(sum_left, zero);
   sum_above = _mm_unpacklo_epi16(sum_above, zero);
   const __m128i sum = _mm_add_epi32(sum_left, sum_above);
-  uint32_t sum32 = _mm_cvtsi128_si32(sum);
+  uint32_t sum32 = (uint32_t)_mm_cvtsi128_si32(sum);
 #if CONFIG_DC_DIV_UNIFY
   int16_t shift = 0;
   uint16_t scale = resolve_divisor_32(24, &shift);
@@ -937,7 +937,7 @@ void aom_highbd_dc_predictor_16x8_sse2(uint16_t *dst, ptrdiff_t stride,
   sum32 += 12;
   sum32 /= 24;
 #endif  // CONFIG_DC_DIV_UNIFY
-  const __m128i row = _mm_set1_epi16((uint16_t)sum32);
+  const __m128i row = _mm_set1_epi16((int16_t)sum32);
   int i;
   for (i = 0; i < 2; ++i) {
     _mm_storeu_si128((__m128i *)dst, row);
@@ -966,7 +966,7 @@ void aom_highbd_dc_predictor_16x32_sse2(uint16_t *dst, ptrdiff_t stride,
   const __m128i zero = _mm_setzero_si128();
   sum_above = _mm_unpacklo_epi16(sum_above, zero);
   const __m128i sum = _mm_add_epi32(sum_left, sum_above);
-  uint32_t sum32 = _mm_cvtsi128_si32(sum);
+  uint32_t sum32 = (uint32_t)_mm_cvtsi128_si32(sum);
 #if CONFIG_DC_DIV_UNIFY
   int16_t shift = 0;
   uint16_t scale = resolve_divisor_32(48, &shift);
@@ -976,7 +976,7 @@ void aom_highbd_dc_predictor_16x32_sse2(uint16_t *dst, ptrdiff_t stride,
   sum32 += 24;
   sum32 /= 48;
 #endif  // CONFIG_DC_DIV_UNIFY
-  const __m128i row = _mm_set1_epi16((uint16_t)sum32);
+  const __m128i row = _mm_set1_epi16((int16_t)sum32);
   int i;
   for (i = 0; i < 8; ++i) {
     _mm_storeu_si128((__m128i *)dst, row);
@@ -1005,7 +1005,7 @@ void aom_highbd_dc_predictor_32x16_sse2(uint16_t *dst, ptrdiff_t stride,
   const __m128i zero = _mm_setzero_si128();
   sum_left = _mm_unpacklo_epi16(sum_left, zero);
   const __m128i sum = _mm_add_epi32(sum_left, sum_above);
-  uint32_t sum32 = _mm_cvtsi128_si32(sum);
+  uint32_t sum32 = (uint32_t)_mm_cvtsi128_si32(sum);
 #if CONFIG_DC_DIV_UNIFY
   int16_t shift = 0;
   uint16_t scale = resolve_divisor_32(48, &shift);
@@ -1015,7 +1015,7 @@ void aom_highbd_dc_predictor_32x16_sse2(uint16_t *dst, ptrdiff_t stride,
   sum32 += 24;
   sum32 /= 48;
 #endif  // CONFIG_DC_DIV_UNIFY
-  const __m128i row = _mm_set1_epi16((uint16_t)sum32);
+  const __m128i row = _mm_set1_epi16((int16_t)sum32);
   int i;
   for (i = 0; i < 4; ++i) {
     _mm_store_si128((__m128i *)dst, row);
